Complete base part HW_02

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepositoryImpl.java
@@ -1,5 +1,7 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
@@ -14,6 +16,7 @@ import java.util.stream.Collectors;
 
 @Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
+    private static final Logger log = LoggerFactory.getLogger(InMemoryMealRepositoryImpl.class);
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
@@ -23,29 +26,44 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     @Override
     public Meal save(Meal meal) {
+        log.info("Save {}",meal);
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
-            repository.put(meal.getId(), meal);
+            repository.put(meal.getId(),meal);
             return meal;
         }
+        if(!repository.containsKey(meal.getId())){
+            return null;
+        }
         // treat case: update, but absent in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        return meal.getUserId().equals(repository.get(meal.getId()).getUserId())
+                ? repository.computeIfPresent(meal.getId(),(id, oldMeal) -> meal) : null;
     }
 
     @Override
     public boolean delete(int id, int userId) {
+        log.info("delete {}", id);
+        if(!repository.containsKey(id)){
+            return false;
+        }
         return repository.get(id).getUserId() == userId && repository.remove(id) != null;
     }
 
     @Override
     public Meal get(int id,int userId) {
+        log.info("get {}" ,id);
+        if(!repository.containsKey(id)){
+            return null;
+        }
         return repository.get(id).getUserId() == userId ? repository.get(id) : null;
     }
 
     @Override
     public Collection<Meal> getAll(int userId) {
+        log.info("getAll() {}" ,userId);
+        System.out.println(repository.values());
         return repository.values().stream()
-                .filter(meal -> meal.getUserId() == userId)
+                .filter(meal -> userId == meal.getUserId())
                 .sorted(Comparator.comparing(Meal::getDateTime).reversed())
                 .collect(Collectors.toList());
     }
